+++ src/ClientApi/Rest/Fetch/Fetch.php
@@ -6,15 +6,16 @@
  * Time: 13:36
  */
 
-namespace Nokaut\ApiKit\ClientApi\Rest\Async;
+namespace Nokaut\ApiKit\ClientApi\Rest\Fetch;
 
 
+use Nokaut\ApiKit\Cache\CacheInterface;
 use Nokaut\ApiKit\ClientApi\Rest\Query\QueryBuilderInterface;
 use Nokaut\ApiKit\Collection\CollectionInterface;
 use Nokaut\ApiKit\Converter\ConverterInterface;
 use Nokaut\ApiKit\Entity\EntityAbstract;
 
-class AsyncFetch
+class Fetch
 {
     /**
      * @var QueryBuilderInterface
@@ -30,13 +31,26 @@ class AsyncFetch
     protected $result;
 
     /**
+     * @var CacheInterface
+     */
+    protected $cache;
+
+    /**
+     * flag show if result was taken form cache/API
+     * @var bool
+     */
+    protected $processed = false;
+
+    /**
      * @param QueryBuilderInterface $query
      * @param ConverterInterface $converter
+     * @param CacheInterface $cache
      */
-    function __construct(QueryBuilderInterface $query, ConverterInterface $converter)
+    function __construct(QueryBuilderInterface $query, ConverterInterface $converter , CacheInterface $cache)
     {
         $this->query = $query;
         $this->converter = $converter;
+        $this->cache = $cache;
     }
 
 
@@ -87,5 +101,46 @@ class AsyncFetch
     {
         return $this->result;
     }
+
+    /**
+     * @param CacheInterface $cache
+     */
+    public function setCache($cache)
+    {
+        $this->cache = $cache;
+    }
+
+    /**
+     * @return CacheInterface
+     */
+    public function getCache()
+    {
+        return $this->cache;
+    }
+
+    /**
+     * @return string
+     */
+    public function prepareCacheKey()
+    {
+        return 'api-' . md5($this->query->createRequestPath());
+    }
+
+    /**
+     * @param boolean $processed
+     */
+    public function setProcessed($processed)
+    {
+        $this->processed = $processed;
+    }
+
+    /**
+     * @return boolean
+     */
+    public function isProcessed()
+    {
+        return $this->processed;
+    }
+
 
 }