+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ChangeLog
 =========
 
-master
+v1.4.0
 ------
  - PhotoUrl - gdy puste photoId zwraca link do zaślepiki noimg<size>.png
  - ProductsRepository - dodanie pola 'click_url' do 'fieldsForProductBox'
@@ -9,6 +9,14 @@ master
  - OffersRepository - funkcja pobierania ofert po shopId
  - OffersRepository - funkcja pobierania oferty po OffersQuery
  - FIX: poprawka do async repository na różne konfiguracje
+
+**UWAGA: zmiana konstruktora w repozytoriach z:**
+
+    public function __construct($apiBaseUrl, ClientApiInterface $clientApi)
+
+**na:**
+
+    public function __construct(Config $config, ClientApiInterface $clientApi)
 
 
 v1.3.1